[Bug-fix] Example script is now correctly dependent on the presence of owntech_examples library.
