+++ owntech/pio_extra.ini
@@ -43,7 +43,7 @@ extra_scripts =
     pre:owntech/scripts/pre_target_gui_config.py
     pre:owntech/scripts/pre_target_install_bootloader.py
     pre:owntech/scripts/pre_bootloader_common.py
-    pre:owntech/scripts/pre_target_samples.py
+    pre:owntech/scripts/pre_target_examples.py
     pre:owntech/scripts/pre_expert_mode.py
     pre:owntech/scripts/pre_stlink.py
     pre:owntech/scripts/middleware_skip_examples.py
@@ -69,7 +69,7 @@ extra_scripts =
     pre:owntech/scripts/pre_target_gui_config.py
     pre:owntech/scripts/pre_bootloader_common.py
     pre:owntech/scripts/pre_bootloader_serial.py
-    pre:owntech/scripts/pre_target_samples.py
+    pre:owntech/scripts/pre_target_examples.py
     pre:owntech/scripts/pre_expert_mode.py
     pre:owntech/scripts/pre_recovery.py
     pre:owntech/scripts/pre_usb.py
