+++ platformio.ini
@@ -46,4 +46,4 @@ monitor_speed = 115200
 # One library per line, 4 spaces at the beginning of the line.
 lib_deps=
     control_pid = https://gitlab.laas.fr/owntech/power-api/opalib-control-pid.git
-    owntech_examples = https://github.com/owntech-foundation/examples.git#0.0.5 
+    owntech_examples = https://github.com/Ayoub-Farah/examples_owntech.git#0.0.6
